+++ stories/react-light-state/utils/mapStateToProps.js
@@ -1,3 +1,0 @@
-export default function mapStateToProps(data, storeName = "lightProps") {
-  return { [storeName]: data };
-}